Closes #1269.

+++ web/src/test/scala/quasar/api/zip.scala
@@ -88,7 +88,7 @@ class ZipSpecs extends Specification with ScalaCheck with ScalazMatchers {
         def loop(acc: ByteVector): ByteVector = {
           val buffer = new Array[Byte](16*1024)
           val n = is.read(buffer)
-          if (n <= 0) acc else loop(acc ++ ByteVector.view(buffer).take(n))
+          if (n <= 0) acc else loop(acc ++ ByteVector.view(buffer).take(n.toLong))
         }
         loop(ByteVector.empty)
       }
@@ -114,7 +114,7 @@ class ZipSpecs extends Specification with ScalaCheck with ScalazMatchers {
     "zip many large files of random bytes (100 MB)" in {
       // NB: this is mainly a performance check. Right now it's about 2 seconds for 100 MB for me.
       val Files = 100
-      val MaxExpectedSize = 1000*1000
+      val MaxExpectedSize = 1000L*1000L
       val MinExpectedSize = MaxExpectedSize / 2
 
       val paths = (0 until Files).toList.map(i => file[Sandboxed]("foo" + i))
@@ -129,7 +129,7 @@ class ZipSpecs extends Specification with ScalaCheck with ScalazMatchers {
       skipped("too slow to run every time (~2 minutes)")
 
       val Files = 10*1000
-      val MaxExpectedSize = 100*1000*1000
+      val MaxExpectedSize = 100L*1000L*1000L
       val MinExpectedSize = MaxExpectedSize / 2
 
       val paths = (0 until Files).toList.map(i => file[Sandboxed]("foo" + i))
